Moved frontend files to the correct directory
