+++ .env
@@ -1,1 +1,1 @@
-VITE_API_URL=http://localhost:8000
+REACT_API_URL=http://localhost:8000
