--- frontend/.env
@@ -1,1 +1,0 @@
-REACT_API_URL=http://localhost:8000
