+++ package.json
@@ -1,22 +1,14 @@
 {
-  "name": "transcript-manager",
+  "name": "frontend",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.14.0",
-    "@emotion/styled": "^11.14.0",
-    "@mui/material": "^6.3.1",
-    "antd": "^5.23.0",
     "axios": "^1.7.9",
     "cra-template": "1.2.0",
-    "file-saver": "^2.0.5",
-    "react": "^19.0.0",
-    "react-csv": "^2.2.2",
-    "react-dom": "^19.0.0",
-    "react-router-dom": "^7.1.3",
-    "react-scripts": "5.0.1",
-    "web-vitals": "^4.2.4",
-    "xlsx": "^0.18.5"
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^7.1.5",
+    "react-scripts": "5.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -41,8 +33,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "@babel/plugin-proposal-private-property-in-object": "^7.21.11"
   }
 }
