+++ src/App.js
@@ -1,15 +1,19 @@
-import React from 'react';
-import AppRouter from './AppRouter';
-import './styles.css';  // 引入样式文件
+import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
+import HomePage from './pages/HomePage';
+import QueryPage from './pages/QueryPage';
+import UploadPage from './pages/UploadPage';
+import './styles/App.css';
 
 function App() {
-    return <AppRouter />;
+  return (
+    <Router>
+      <Routes>
+        <Route path="/" element={<HomePage />} />
+        <Route path="/upload" element={<UploadPage />} />
+        <Route path="/query" element={<QueryPage />} />
+      </Routes>
+    </Router>
+  );
 }
 
 export default App;
-
-
-
-
-
-
