+++ src/pages/QueryPage.js
@@ -1,125 +1,103 @@
-import React, { useCallback, useMemo, useState, useEffect } from 'react';
-import { CSVLink } from 'react-csv';
-import * as XLSX from 'xlsx';
+import React, { useCallback, useState} from 'react';
+import { searchData } from '../apiServices'
 import NotificationBar from '../components/NotificationBar';
-import TableComponent from '../components/TableComponent';
-import SearchComponent from '../components/SearchComponent';
-import { fetchExistingData, querySearchData } from '../api';
-import '../styles.css';
-
-const categoryList = [
-    'Accounting', 'Communication', 'Curriculum', 'eBiz', 'Economics', 'Educ Subjects',
-    'Education Leadership', 'Ethics', 'Finance', 'HR/ORG/Leader', 'International Business',
-    'IT General', 'Law', 'Management', 'Marketing', 'Operations Mgmt',
-    'Other Education Leadership', 'Project Mgmt', 'Research/Stats', 'Strategy'
-];
-
-const educationLevels = ['Bachelor', 'Master', 'Doctorate'];
-
-const ExportButtons = React.memo(({ data }) => {
-    const headers = useMemo(() => [
-        { label: 'Educator', key: 'educatorname' },
-        { label: 'Course Name', key: 'course_name' },
-        { label: 'Credits Earned', key: 'credits_earned' },
-        { label: 'Grade', key: 'grade' },
-        { label: 'Category', key: 'should_be_category' }
-    ], []);
-
-    const exportToExcel = useCallback(() => {
-        if (!data || data.length === 0) return;
-        const worksheet = XLSX.utils.json_to_sheet(data);
-        const workbook = XLSX.utils.book_new();
-        XLSX.utils.book_append_sheet(workbook, worksheet, 'Transcripts');
-        XLSX.writeFile(workbook, 'exported_data.xlsx');
-    }, [data]);
-
-    if (!data || data.length === 0) return null;
-
-    return (
-        <div className="export-container">
-            <CSVLink data={data} headers={headers} filename="exported_data.csv" className="export-btn">
-                Export as CSV
-            </CSVLink>
-            <button className="export-btn" onClick={exportToExcel}>Export as Excel</button>
-        </div>
-    );
-});
+import SearchCriteria from '../components/SearchCriteria';
+import ResultTable from '../components/ResultTable';
+import '../styles/QueryPage.css';
 
 const QueryPage = () => {
-    const [data, setData] = useState([]);
-    const [searchQuery, setSearchQuery] = useState('');
-    const [selectedCategory, setSelectedCategory] = useState('');
-    const [selectedEducation, setSelectedEducation] = useState([]);
-    const [notification, setNotification] = useState('');
-    const [notificationType, setNotificationType] = useState('');
-    const [searched, setSearched] = useState(false);
+  const [criteria, setCriteria] = useState({ 
+    name: '',
+    courseCategory: '',
+    educationLevel: [],
+  }); // State for search criteria
+  const [results, setResults] = useState([]); // State to store search results
+  const [notification, setNotification] = useState('');
+  const [notificationType, setNotificationType] = useState(null);
 
-    useEffect(() => {
-        fetchExistingData()
-            .then(result => {
-                setData(result);
-                setNotification('Data loaded successfully.');
-                setNotificationType('success');
-            })
-            .catch(() => {
-                setNotification('Error loading data.');
-                setNotificationType('error');
-            });
-    }, []);
+  // Handles the search request
+  const handleSearch = useCallback(async () => {
+    if (!criteria.name.trim() && !criteria.courseCategory && criteria.educationLevel.length === 0) {
+      setResults([]); // Clear previous results if no criteria is entered
+      return;
+    }
 
-    const handleSearch = useCallback(async () => {
-        if (!searchQuery.trim() && !selectedCategory && selectedEducation.length === 0) {
-            setData([]);
-            setSearched(false);
-            return;
-        }
-        try {
-            const searchParams = { query: searchQuery, category: selectedCategory, education: selectedEducation };
-            const filteredData = await querySearchData(searchParams);
-            setData(filteredData);
-            setNotification('Search completed successfully.');
-            setNotificationType('success');
-            setSearched(true);
-        } catch (error) {
-            setNotification('Error during search.');
-            setNotificationType('error');
-        }
-    }, [searchQuery, selectedCategory, selectedEducation]);
+    try {
+      const searchResults = await searchData(criteria); // Call API with criteria
+      setResults(searchResults); // Store the response data
+      setNotification('Search completed successfully.');
+      setNotificationType('success');
+    } catch (error) {
+      setNotification('Error during search.');
+      setNotificationType('error');
+    }
+  }, [criteria]);
 
-    return (
-        <div className="container query-page">
-            <h1>Query and Download</h1>
-            <NotificationBar message={notification} type={notificationType} />
-            <SearchComponent
-                searchQuery={searchQuery}
-                setSearchQuery={setSearchQuery}
-                selectedCategory={selectedCategory}
-                setSelectedCategory={setSelectedCategory}
-                selectedEducation={selectedEducation}
-                setSelectedEducation={setSelectedEducation}
-                handleSearch={handleSearch}
-                categoryList={categoryList}
-                educationLevels={educationLevels}
-            />
-            {searched && data.length > 0 && (
-                <>
-                    <TableComponent data={data} />
-                    <ExportButtons data={data} />
-                </>
-            )}
-        </div>
-    );
-};
+  // Handles the download request
+  const handleDownload = () => {
+    if (!results.queried_data || results.queried_data.length === 0) {
+      setNotification('No data available to download.');
+      setNotificationType('error');
+      return;
+    }
+
+    // Extract educator's name and format a filename
+    const name = results.educator_name;
+    const fileName = `${name.replace(/\s+/g, '_')}_Qualifications_Worksheet.csv`;
+
+    // Convert search results to CSV
+    const csvHeaders = `"Faculty's Name","${name}"\n`; 
+    const csvRows = results.queried_data.map(row => 
+      Object.values(row).map(value => `"${value}"`).join(",") // Ensure CSV-safe formatting
+    ).join("\n");
+
+    const csvContent = csvHeaders + csvRows;
+
+    // Create a Blob and download the CSV
+    const blob = new Blob([csvContent], { type: 'text/csv' });
+    const url = window.URL.createObjectURL(blob);
+
+    // Create and trigger a download link
+    const link = document.createElement('a');
+    link.href = url;
+    link.setAttribute('download', fileName);
+    document.body.appendChild(link);
+    link.click();
+
+    // Clean up by removing the link
+    document.body.removeChild(link);
+    window.URL.revokeObjectURL(url); // Free up memory
+
+    setNotification('CSV downloaded successfully.');
+    setNotificationType('success');
+  };
+
+  return (
+    <div className="container">
+      <h1>Query and Download Data</h1>
+      {/* Notification Bar */}
+      {notification && notificationType && (
+        <NotificationBar message={notification} type={notificationType} />
+      )}
+
+      {/* Search Criteria Component */}
+      <SearchCriteria 
+        criteria={criteria} 
+        setCriteria={setCriteria} 
+        handleSearch={handleSearch} 
+      />
+
+      {/* Display ResultTable and export button only if results exist */}
+      {results.queried_data && results.queried_data.length > 0 && (
+        <>
+          <ResultTable data={results} />
+          <button className="export-btn" onClick={handleDownload}>
+            Export as CSV
+          </button>
+        </>
+      )}
+    </div>
+  );
+}
 
 export default QueryPage;
-
-
-
-
-
-
-
-
-
-
-
