--- src/styles.css
@@ -1,324 +1,0 @@
-/* 全局样式 */
-body {
-    font-family: 'Arial', sans-serif;
-    margin: 0;
-    padding: 0;
-    background-color: #f4f7f6; 
-    display: flex;
-    justify-content: center; 
-    align-items: flex-start; 
-    height: 100vh; 
-    padding: 0 20px; 
-}
-
-/* 主页特定样式 */
-.container {
-    width: 100%; 
-    max-width: 700px; 
-    height: auto;
-    padding: 20px; 
-    box-sizing: border-box;
-    background-color: transparent; 
-    margin-top: 80px; 
-    text-align: center; 
-}
-
-/* 头部样式 */
-h1 {
-    font-size: 2.5em;
-    color: #333;
-    margin-top: 10px; 
-    margin-bottom: 10px; 
-    font-weight: bold;
-}
-
-/* 主页副标题样式 */
-p {
-    font-size: 1.3em; 
-    color: rgba(51, 51, 51, 0.7); 
-    margin-bottom: 15px; 
-}
-
-/* 按钮容器样式 */
-.button-container {
-    display: flex;
-    flex-direction: column;  
-    align-items: center; 
-    gap: 20px; 
-    width: 100%; 
-    margin-top: 150px; 
-}
-
-/* 按钮样式 */
-button {
-    padding: 24px 100px; 
-    font-size: 1.4em;
-    border-radius: 5px;
-    cursor: pointer;
-    transition: all 0.3s ease;
-    width: 100%; 
-    max-width: 500px; 
-    min-width: 500px; 
-    border: none; 
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1); 
-}
-
-/* 主按钮样式 */
-button.primary-btn {
-    background-color: #2e7ed3;
-    color: white;
-}
-
-button.primary-btn:hover {
-    background-color: #37679b;
-    transform: scale(1.05);
-    box-shadow: 0 6px 12px rgba(0, 0, 0, 0.2); 
-}
-
-/* 次按钮样式 */
-button.secondary-btn {
-    background-color: #2ba94a;
-    color: white;
-}
-
-button.secondary-btn:hover {
-    background-color: #30c04f;
-    transform: scale(1.05);
-    box-shadow: 0 6px 12px rgba(0, 0, 0, 0.2); 
-}
-
-/* 导出按钮容器 - */
-.export-container {
-    display: flex;
-    justify-content: center; 
-    gap: 20px; 
-    margin-top: 20px;
-}
-
-/* 导出按钮样式 */
-.export-btn {
-    display: inline-block; 
-    padding: 10px 20px; 
-    background-color: #2ba94a; 
-    color: white; 
-    font-size: 16px; 
-    border: none; 
-    border-radius: 5px; 
-    cursor: pointer; 
-    transition: all 0.3s ease; 
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1); 
-    text-align: center; 
-    width: auto; 
-    min-width: fit-content; 
-    max-width: 180px; 
-    box-sizing: border-box; 
-}
-
-/* 悬停效果 */
-.export-btn:hover {
-    background-color: #30c04f; 
-    transform: scale(1.05); 
-    box-shadow: 0 6px 12px rgba(0, 0, 0, 0.2); 
-}
-
-/* 搜索框 + 按钮 + 下拉菜单 */
-.search-box {
-    display: flex;
-    justify-content: center; 
-    align-items: center;
-    gap: 15px; 
-    margin-bottom: 15px; 
-}
-
-/* 搜索输入框 */
-.search-input {
-    width: 300px;
-    padding: 10px;
-    font-size: 16px;
-    border: 2px solid #ddd;
-    border-radius: 5px;
-}
-
-/* 搜索按钮 */
-.search-btn {
-    padding: 10px 20px; 
-    font-size: 16px; 
-    border: none; 
-    border-radius: 4px; 
-    background-color: #2e7ed3; 
-    color: white; 
-    cursor: pointer; 
-    transition: background-color 0.2s ease; 
-    width: auto; 
-    min-width: 80px; 
-}
-
-/* 悬停效果 */
-.search-btn:hover {
-    background-color: #37679b; 
-}
-
-/* 下拉菜单容器 */
-.dropdown-container {
-    margin-top: 10px; 
-}
-
-/* 下拉菜单 */
-.dropdown-select {
-    width: 180px;
-    padding: 10px;
-    font-size: 16px;
-    border: 2px solid #ddd;
-    border-radius: 4px;
-    background-color: white;
-    cursor: pointer;
-}
-/* CheckBox 组：让其在搜索框下方，并水平排列 */
-.checkbox-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin-top: 10px;
-}
-
-/* CheckBox 组标题 */
-.checkbox-title {
-    font-size: 16px;
-    font-weight: bold;
-    margin-bottom: 5px;
-}
-
-/* CheckBox 组的排列方式 */
-.checkbox-group {
-    display: flex;
-    justify-content: center;
-    gap: 20px; 
-    flex-wrap: wrap; 
-}
-
-/* CheckBox 标签 */
-.checkbox-label {
-    font-size: 14px;
-    display: flex;
-    align-items: center;
-    gap: 5px;
-}
-
-/* 上传页面容器样式 */
-.container {
-    text-align: center;
-    margin: 50px auto;
-    max-width: 600px;
-    background-color: #f9f9f9;
-    padding: 20px;
-    border-radius: 8px;
-    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
-}
-
-/* 标题样式 */
-.container h1 {
-    font-size: 2em;
-    color: #333;
-    margin-bottom: 15px;
-}
-
-/* 描述文字样式 */
-.container p {
-    font-size: 1em;
-    color: #555;
-    margin-bottom: 20px;
-    line-height: 1.5;
-}
-
-/* 文件输入按钮样式 */
-input[type="file"] {
-    appearance: none; 
-    background-color: #2e7ed3; 
-    color: rgb(252, 250, 250); 
-    font-size: 16px; 
-    padding: 80px 100px; 
-    border: none; 
-    border-radius: 5px; 
-    cursor: pointer; 
-    transition: background-color 0.3s ease, transform 0.2s ease; 
-}
-
-/* 鼠标悬停效果 */
-input[type="file"]:hover {
-    background-color: #669fdc; 
-    transform: scale(1.05); 
-}
-
-/* 鼠标按下效果 */
-input[type="file"]:active {
-    background-color: #1f4d6b; 
-    transform: scale(0.95); 
-}
-
-
-input[type="file"]:hover {
-    border-color: #2e7ed3;
-}
-
-/* 成功消息样式 */
-.success-text {
-    color: #4CAF50;
-    font-size: 1em;
-    margin-top: 15px;
-}
-
-/* 错误消息样式 */
-.error-text {
-    color: #E53935;
-    font-size: 1em;
-    margin-top: 15px;
-    white-space: pre-wrap; 
-}
-
-/* 上传文件列表 */
-.container ul {
-    list-style-type: none;
-    padding: 0;
-    margin-top: 20px;
-}
-
-.container ul li {
-    padding: 8px;
-    font-size: 0.95em;
-    border-bottom: 1px solid #ddd;
-}
-
-/* 鼠标悬停文件列表高亮 */
-.container ul li:hover {
-    background-color: #f4f4f4;
-}
-
-/* 上传文件标题样式 */
-.container h3 {
-    font-size: 1.2em;
-    color: #333;
-    margin-bottom: 10px;
-}
-
-
-/* 进度条容器 */
-.progress-bar-container {
-    width: 100%;
-    background-color: #ddd;
-    border-radius: 5px;
-    overflow: hidden;
-    margin-top: 5px;
-}
-
-/* 进度条 */
-.progress-bar {
-    height: 10px;
-    background-color: #2ba94a; 
-    text-align: center;
-    line-height: 10px;
-    color: white;
-    font-size: 10px;
-    border-radius: 5px;
-    transition: width 0.3s ease-in-out;
-}
-
