+++ src/styles/NotificationBar.css
