+++ src/styles/QueryPage.css
