+++ src/styles/SearchCriteria.css
